+++ R/imports-base.R
@@ -5,4 +5,7 @@
 # methods ####
 #' @importFrom methods formalArgs
 #' @importFrom methods functionBody
+
+# utils ####
+#' @importFrom utils getS3method
 NULL